personal_projects/31.ai-youtube-analysis/requirements.txt
pytube 
controlflow 
openai 
python-dotenv
google-api-python-client
# langchain_experimental.